Ulepszanie menu. Dodano prawidlowe dzialanie opcji resume. Klasa leveling powoli zaczyna sie szykowac do obslugi wyboru levelu. Poprawiono bledy w engine::nextlevel, oraz engine::load. Wystepuja blad zwiazany z obsluga kolizji miedzy playerem a jednostkami- w 1 levelu nie wykrywane sa te kolizje. Problem pominiety narazie (prawdopodobnie w 1 levelu nie bedzie jednostek)

+++ src/Level.cpp
@@ -23,14 +23,13 @@ void Level::LoadFromFile( const string& FileName){
     
   int EntityAmount=0;  
   lvl>> EntityAmount;
-    //m_entityF.reset(new EntityFactory());
   
   
-  for (size_t i=0; i<EntityAmount; ++i){
+  for (ushort i=0; i<EntityAmount; ++i){
     ushort x=0,y=0;
     string name;
     lvl>>x>>y>>name;
-   ( m_entityF)->AddEntity(x,y,name);
+   ( m_entityF )->AddEntity(x,y,name);
   }  
     
 }
